refactor(platform): brand 아이콘 도입 + Label 단일 일원화 + UX 다듬기
브랜드 아이콘 (@icons-pack/react-simple-icons 추가):
- JiraConnectForm EmptyState: lucide Link → SiJirasoftware (Atlassian blue)
- GithubConnectForm EmptyState: lucide Github → SiGithub (브랜드 검정 + dark:invert)
- SettingsTab의 [Jira] / [GitHub] sub-tab trigger에 brand 아이콘 추가
- color="default" + simple-icons 컬러 사용. 향후 플랫폼 추가 시 동일 패턴

라벨 다중 → 단일 일원화:
- 신규 LabelCombobox.tsx (단일 선택, Repo 콤보 패턴 + 색상 dot)
- 기존 LabelMultiSelect.tsx / AssigneeMultiSelect.tsx 삭제 (사용처 없음)
- GithubConnectForm DefaultIssueSettingsFields는 LabelCombobox 사용 (defaults.labels는 string[] 유지 — 단일 항목 array)
- Assignee는 default 설정에서 빠져 있음 (등록 시점에 선택 — Jira와 일관)

UX 다듬기:
- Empty state wrapping div(`mt-3 w-full max-w-[260px]`) 양쪽 제거 — 빈 reserved 공간 군더더기 제거. Banner 직접 conditional render
- github.patButton: "PAT" → "PAT 인증" (약어만 두면 의미 약함, 동작 의도 명확화)
- GitHub viewer 카드 desc는 viewerEmail 우선, 없으면 "github.com" fallback (재로그인해야 user:email scope 적용 → email 채워짐)

pnpm typecheck/test 그린, pnpm build 그린.
